+++ src/main/java/agents/flexibility/dynamicProgramming/states/EnergyStateManager.java
@@ -4,7 +4,6 @@
 package agents.flexibility.dynamicProgramming.states;
 
 import java.util.ArrayList;
-import java.util.stream.IntStream;
 import agents.flexibility.GenericDevice;
 import agents.flexibility.GenericDeviceCache;
 import agents.flexibility.dynamicProgramming.Optimiser;
@@ -16,6 +15,9 @@ import de.dlr.gitlab.fame.time.TimeStamp;
  * 
  * @author Christoph Schimeczek, Felix Nitsch, Johannes Kochems */
 public class EnergyStateManager implements StateManager {
+	/** Added to floating point calculation of transition steps to avoid rounding errors */
+	private static final double PRECISION_GUARD = 1E-5;
+
 	private final GenericDevice device;
 	private final GenericDeviceCache deviceCache;
 	private final AssessmentFunction assessmentFunction;
@@ -31,8 +33,8 @@ public class EnergyStateManager implements StateManager {
 	private int currentOptimisationTimeIndex;
 
 	private boolean hasSelfDischarge;
-	private double[] energyDeltasCharging;
-	private double[] energyDeltasDischarging;
+	private double[] transitionValuesCharging;
+	private double[] transitionValuesDischarging;
 
 	public EnergyStateManager(GenericDevice device, AssessmentFunction assessmentFunction, double planningHorizonInHours,
 			double energyResolutionInMWH) {
@@ -94,76 +96,76 @@ public class EnergyStateManager implements StateManager {
 		currentOptimisationTimeIndex = (int) ((time.getStep() - startingPeriod.getStartTime().getStep())
 				/ startingPeriod.getDuration().getSteps());
 		if (!hasSelfDischarge) {
-			cacheEnergyDeltas();
+			cacheTransitionValuesNoSelfDischarge();
 		}
 	}
 
-	/** Cache deltas between energy levels - only applicable without self discharge */
-	private void cacheEnergyDeltas() {
-		int maxChargingSteps = (int) Math.floor(deviceCache.getMaxNetChargingEnergyInMWH() / energyResolutionInMWH);
-		energyDeltasCharging = new double[maxChargingSteps + 1];
+	/** Cache values of transitions - only applicable without self discharge */
+	private void cacheTransitionValuesNoSelfDischarge() {
+		int maxChargingSteps = (int) Math
+				.floor(deviceCache.getMaxNetChargingEnergyInMWH() / energyResolutionInMWH + PRECISION_GUARD);
+		transitionValuesCharging = new double[maxChargingSteps + 1];
 		for (int chargingSteps = 0; chargingSteps <= maxChargingSteps; chargingSteps++) {
-			energyDeltasCharging[chargingSteps] = deviceCache.simulateTransition(0, chargingSteps * energyResolutionInMWH);
+			transitionValuesCharging[chargingSteps] = calcValueFor(0, chargingSteps);
 		}
-		int maxDischargingSteps = -(int) Math.ceil(deviceCache.getMaxNetDischargingEnergyInMWH() / energyResolutionInMWH);
-		energyDeltasDischarging = new double[maxDischargingSteps + 1];
+		int maxDischargingSteps = -(int) Math
+				.ceil(deviceCache.getMaxNetDischargingEnergyInMWH() / energyResolutionInMWH - PRECISION_GUARD);
+		transitionValuesDischarging = new double[maxDischargingSteps + 1];
 		for (int dischargingSteps = 0; dischargingSteps <= maxDischargingSteps; dischargingSteps++) {
-			energyDeltasDischarging[dischargingSteps] = deviceCache.simulateTransition(0,
-					-dischargingSteps * energyResolutionInMWH);
+			transitionValuesDischarging[dischargingSteps] = calcValueFor(0, -dischargingSteps);
 		}
 	}
 
 	@Override
+	public boolean useStateList() {
+		return false;
+	}
+
+	@Override
 	public int[] getInitialStates() {
-		int lowestIndex = energyToCeilIndex(deviceCache.getEnergyContentLowerLimitInMWH());
-		int highestIndex = energyToFloorIndex(deviceCache.getEnergyContentUpperLimitInMWH());
-		return IntStream.range(lowestIndex, highestIndex + 1).toArray();
+		final int lowestIndex = energyToCeilIndex(deviceCache.getEnergyContentLowerLimitInMWH());
+		final int highestIndex = energyToFloorIndex(deviceCache.getEnergyContentUpperLimitInMWH());
+		return new int[] {lowestIndex, highestIndex};
 	}
 
 	/** @return next lower index corresponding to given energy level */
 	private int energyToFloorIndex(double energyAmountInMWH) {
-		double energyLevel = Math.floor(energyAmountInMWH / energyResolutionInMWH) * energyResolutionInMWH;
+		double energyLevel = Math.floor(energyAmountInMWH / energyResolutionInMWH + PRECISION_GUARD)
+				* energyResolutionInMWH;
 		return (int) Math.round((energyLevel - lowestLevelEnergyInMWH) / energyResolutionInMWH);
 	}
 
 	/** @return next higher index corresponding to given energy level */
 	private int energyToCeilIndex(double energyAmountInMWH) {
-		double energyLevel = Math.ceil(energyAmountInMWH / energyResolutionInMWH) * energyResolutionInMWH;
+		double energyLevel = Math.ceil(energyAmountInMWH / energyResolutionInMWH - PRECISION_GUARD) * energyResolutionInMWH;
 		return (int) Math.round((energyLevel - lowestLevelEnergyInMWH) / energyResolutionInMWH);
 	}
 
 	@Override
 	public int[] getFinalStates(int initialStateIndex) {
-		double initialEnergyContentInMWH = initialStateIndex * energyResolutionInMWH;
-		double lowestEnergyContentInMWH = deviceCache.getMinTargetEnergyContentInMWH(initialEnergyContentInMWH);
-		double highestEnergyContentInMWH = deviceCache.getMaxTargetEnergyContentInMWH(initialEnergyContentInMWH);
-		return IntStream
-				.range(energyToCeilIndex(lowestEnergyContentInMWH), energyToFloorIndex(highestEnergyContentInMWH) + 1)
-				.toArray();
+		final double initialEnergyContentInMWH = initialStateIndex * energyResolutionInMWH;
+		final double lowestEnergyContentInMWH = deviceCache.getMinTargetEnergyContentInMWH(initialEnergyContentInMWH);
+		final double highestEnergyContentInMWH = deviceCache.getMaxTargetEnergyContentInMWH(initialEnergyContentInMWH);
+		return new int[] {energyToCeilIndex(lowestEnergyContentInMWH), energyToFloorIndex(highestEnergyContentInMWH)};
 	}
 
 	@Override
 	public double getTransitionValueFor(int initialStateIndex, int finalStateIndex) {
-		// TODO: cache whole transition if possible, including assessment!
-		double externalEnergyDeltaInMWH = getEnergyDeltaInMWH(initialStateIndex, finalStateIndex);
-		return assessmentFunction.assessTransition(externalEnergyDeltaInMWH);
+		return hasSelfDischarge ? calcValueFor(initialStateIndex, finalStateIndex)
+				: getCachedValueFor(initialStateIndex, finalStateIndex);
 	}
 
-	/** @return energy delta corresponding to given indices of initial and final state */
-	private double getEnergyDeltaInMWH(int initialStateIndex, int finalStateIndex) {
-		if (!hasSelfDischarge) {
-			return getCachedEnergyDeltaInMWH(initialStateIndex, finalStateIndex);
-		}
-		return getSimulatedEnergyDeltaInMWH(initialStateIndex, finalStateIndex);
-	}
-
-	private double getCachedEnergyDeltaInMWH(int initialStateIndex, int finalStateIndex) {
+	/** @return cached value of transition, only available without self discharge */
+	private double getCachedValueFor(int initialStateIndex, int finalStateIndex) {
 		int stateDelta = finalStateIndex - initialStateIndex;
-		return stateDelta >= 0 ? energyDeltasCharging[stateDelta] : energyDeltasDischarging[-stateDelta];
+		return stateDelta >= 0 ? transitionValuesCharging[stateDelta] : transitionValuesDischarging[-stateDelta];
 	}
 
-	private double getSimulatedEnergyDeltaInMWH(int initialStateIndex, int finalStateIndex) {
-		return deviceCache.simulateTransition(indexToEnergy(initialStateIndex), indexToEnergy(finalStateIndex));
+	/** @return calculated value of transition */
+	private double calcValueFor(int initialStateIndex, int finalStateIndex) {
+		double externalEnergyDeltaInMWH = deviceCache.simulateTransition(indexToEnergy(initialStateIndex),
+				indexToEnergy(finalStateIndex));
+		return assessmentFunction.assessTransition(externalEnergyDeltaInMWH);
 	}
 
 	/** @return energy content corresponding to the given index */
@@ -172,14 +174,16 @@ public class EnergyStateManager implements StateManager {
 	}
 
 	@Override
-	public double getBestValueNextPeriod(int finalStateIndex) {
-		return currentOptimisationTimeIndex + 1 < numberOfTimeSteps
-				? bestValue[currentOptimisationTimeIndex + 1][finalStateIndex]
-				: getWaterValue();
-	}
-
-	private double getWaterValue() {
-		return 0; // TODO: implement
+	public double[] getBestValuesNextPeriod() {
+		if (currentOptimisationTimeIndex + 1 < numberOfTimeSteps) {
+			return bestValue[currentOptimisationTimeIndex + 1];
+		} else {
+			double[] bestValues = new double[numberOfEnergyStates];
+			for (int index = 0; index < numberOfEnergyStates; index++) {
+				bestValues[index] = 0;
+			}
+			return bestValues;
+		}
 	}
 
 	@Override